[REFACTOR] Moved everything related to the event organization to a package
