+++ Eventy/app/src/main/java/com/example/eventy/events/organizeevent/EventAgendaCreation.java
@@ -1,4 +1,4 @@
-package com.example.eventy.events;
+package com.example.eventy.events.organizeevent;
 
 import android.os.Bundle;
