+++ Eventy/app/src/main/res/layout/fragment_event_agenda_creation.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".events.EventAgendaCreation"
+    tools:context=".events.organizeevent.EventAgendaCreation"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_marginTop="5dp"
