+++ Eventy/app/src/main/res/layout/fragment_event_invitation_sending.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background_color"
-    tools:context=".events.EventInvitationSendingFragment">
+    tools:context=".events.organizeevent.EventInvitationSendingFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
