+++ Eventy/app/src/main/res/layout/fragment_event_organization.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:fillViewport="true"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".events.EventOrganizationFragment">
+    tools:context=".events.organizeevent.EventOrganizationFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
