+++ Eventy/app/src/main/res/layout/fragment_event_organization_basic_information.xml
@@ -2,7 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".events.EventOrganizationBasicInformationFragment"
+    tools:context=".events.organizeevent.EventOrganizationBasicInformationFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
