+++ Eventy/app/src/main/res/layout/fragment_see_agenda.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".events.SeeAgendaFragment">
+    tools:context=".events.organizeevent.SeeAgendaFragment">
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerView"
         android:layout_width="match_parent"
