+++ Eventy/app/src/main/res/navigation/mobile_navigation.xml
@@ -37,7 +37,7 @@
 
     <fragment
         android:id="@+id/nav_event_organization"
-        android:name="com.example.eventy.events.EventOrganizationFragment"
+        android:name="com.example.eventy.events.organizeevent.EventOrganizationFragment"
         android:label="Organize Event"
         tools:layout="@layout/fragment_event_organization"/>
